+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/auth/navigation/AuthNavGraph.kt
@@ -3,6 +3,8 @@ package com.example.seollyongdan_frontend.presentation.auth.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.example.seollyongdan_frontend.presentation.auth.screen.SignUpRoute
+import com.example.seollyongdan_frontend.presentation.auth.screen.LoginRoute
+
 
 fun NavGraphBuilder.loginNavGraph(
     navigator: AuthNavigator
