+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/community/navigation/CommunityNavGraph.kt
@@ -2,6 +2,7 @@ package com.example.seollyongdan_frontend.presentation.community.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.example.seollyongdan_frontend.presentation.community.screen.CommunityRoute
 
 
 fun NavGraphBuilder.communityNavGraph(
