+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/guide/navigation/GuideNavGraph.kt
@@ -2,6 +2,7 @@ package com.example.seollyongdan_frontend.presentation.guide.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.example.seollyongdan_frontend.presentation.guide.screen.GuideRoute
 
 fun NavGraphBuilder.guideNavGraph(
     navigator: GuideNavigator
