+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/guide/screen/GuideScreen.kt
@@ -1,9 +1,10 @@
-package com.example.seollyongdan_frontend.presentation.guide.navigation
+package com.example.seollyongdan_frontend.presentation.guide.screen
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.graphics.Color
+import com.example.seollyongdan_frontend.presentation.guide.navigation.GuideNavigator
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 @Composable
