+++ app/src/main/java/com/example/seollyongdan_frontend/presentation/mypage/navigation/MypageNavGraph.kt
@@ -1,9 +1,14 @@
 package com.example.seollyongdan_frontend.presentation.mypage.navigation
 
 import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.example.seollyongdan_frontend.presentation.mypage.screen.MypageRoute
 
 fun NavGraphBuilder.mypageNavGraph(
     navigator: MypageNavigator
 ){
+    composable(route = "home"){
+        MypageRoute(navigator = navigator)
+    }
 
 }
