+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Sellyongdan-frontend</string>
+    <string name="app_name">Seollyongdan-frontend</string>
 </resources>
