+++ app/src/main/res/values/themes.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.Sellyongdanfrontend" parent="android:Theme.Material.Light.NoActionBar" />
+    <style name="Theme.Seollyongdanfrontend" parent="android:Theme.Material.Light.NoActionBar" />
 </resources>
